#41 introduced DataManagerException

+++ src/main/java/ru/zhelper/zhelper/services/ProcurementDataManagerImpl.java
@@ -4,8 +4,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.persistence.EntityNotFoundException;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,42 +11,76 @@ import org.springframework.stereotype.Service;
 
 import ru.zhelper.zhelper.models.Procurement;
 import ru.zhelper.zhelper.repository.ProcurementRepo;
+import ru.zhelper.zhelper.services.exceptions.DataManagerException;
 
 @Service
 public class ProcurementDataManagerImpl implements ProcurementDataManager {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(ProcurementDataManagerImpl.class);
-
-	private static final String NO_PROCUREMENT_FOUND_MESSAGE = "Found no Procurement with id {}!";
 	
     @Autowired
 	private ProcurementRepo repository;
 
 	@Override
-	public Procurement loadEntity(Long id) {
-		Procurement procurement = null;
+	public Procurement loadEntity(Long idToLoad) {
+		if (idToLoad == null) {
+			throw new DataManagerException(
+					DataManagerException.COULD_NOT_LOAD_PROCUREMENT_NULL_DATA, null);
+		}
 		
+		Procurement procurement = null;
 		try {
-			procurement = repository.getById(id);
-		} catch (EntityNotFoundException e) {
-			LOGGER.warn(NO_PROCUREMENT_FOUND_MESSAGE, id);
+			procurement = repository.getById(idToLoad);
+		} catch (RuntimeException e) {
+			String msg = String.format(
+					DataManagerException.NON_EXISTING_LOAD_OR_DELETE_EXCEPTION, idToLoad);
+			LOGGER.warn(msg);
+			throw new DataManagerException(msg, e);
 		}
 		return procurement;
 	}
 
 	@Override
 	public Procurement saveEntity(Procurement procurement) {
-		return repository.save(procurement);
+		if (procurement == null) {
+			throw new DataManagerException(
+					DataManagerException.COULD_NOT_SAVE_PROCUREMENT_NULL_DATA, null);
+		}
+
+		try {
+			return repository.save(procurement);
+		} catch (DataManagerException dataMgrExc) {
+			LOGGER.warn(DataManagerException.COULD_NOT_SAVE_PROCUREMENT);
+			throw new DataManagerException(DataManagerException.COULD_NOT_SAVE_PROCUREMENT, dataMgrExc);
+		}
 	}
 
 	@Override
 	public void deleteEntity(Procurement procurement) {
-		repository.delete(procurement);
+		if (procurement == null) {
+			throw new DataManagerException(
+					DataManagerException.COULD_NOT_DELETE_PROCUREMENT_NULL_DATA, null);
+		}
+		try {
+			repository.delete(procurement);
+		} catch (RuntimeException e) {
+			String msg = String.format(
+					DataManagerException.NON_EXISTING_LOAD_OR_DELETE_EXCEPTION, procurement.getId());
+			LOGGER.warn(msg);
+			throw new DataManagerException(msg, e);
+		}
 	}
 
 	@Override
 	public void deleteEntityById(Long idToDelete) {
-		repository.deleteById(idToDelete);
+		try {
+			repository.deleteById(idToDelete);
+		} catch (RuntimeException e) {
+			String msg = String.format(
+					DataManagerException.NON_EXISTING_LOAD_OR_DELETE_EXCEPTION, idToDelete);
+			LOGGER.warn(msg);
+			throw new DataManagerException(msg, e);
+		}
 	}
 	
 	@Override
